Avoid calling the done callback twice.
Previously, this code would call the passed in done once when
`clang-format` exited because we passed it down into
`actualClangFormat`, and once when the stream ended, because we
explicitly called it.

It is a bit unclear whose responsibility it should be to call the
callback. However because only `actualClangFormat` knows the exit error
code, we must handle the call there, which then means we must remove the
"end" handler here.